break down reserves utils into small modules
