+++ src/utils/assetPrices.ts
@@ -1,9 +1,8 @@
-import { parsePriceData, PythHttpClient } from "@pythnetwork/client";
-import SwitchboardProgram from "@switchboard-xyz/sbv2-lite/";
-import { AggregatorState } from "@switchboard-xyz/switchboard-api/lib/compiled";
 import { AccountInfo, PublicKey } from "@solana/web3.js";
+import { parsePriceData } from "@pythnetwork/client";
+import SwitchboardProgram from "@switchboard-xyz/sbv2-lite";
+import { AggregatorState } from "@switchboard-xyz/switchboard-api/lib/compiled";
 import { CONNECTION } from "common/config";
-
 
 const connection = CONNECTION;
 const SBV1_MAINNET = 'DtmE9D2CSB4L5D6A15mraeEjrGMm6auWVzgaD8hK2tZM';
@@ -21,7 +20,39 @@ export const getOracleAddresses = (parsedReserves: ParsedReserve[]) => {
     return oracles;
 }
 
-export const getPythAccountsInfo = async (oracles: Map<string, { pyth: string, sb: string }>) => {
+export const getAssetPrices = async (oracles: Map<string, { pyth: string; sb: string; }>, sbv2Program: SwitchboardProgram) => {
+    const prices = new Map<string, number>();
+    const pythAccountsInfo = await getPythAccountsInfo(oracles);
+    const sbAccountsInfo = await getSbAccountsInfo(oracles);
+
+    for (const [reserveAddress, { pyth, sb }] of oracles) {
+        if (pyth) {
+            try {
+                const pythAccountInfo = pythAccountsInfo.get(reserveAddress);
+                if (pythAccountInfo) {
+                    const price = getPriceFromPyth(pythAccountInfo);
+                    prices.set(reserveAddress, price);
+                    continue;
+                }
+            } catch { }
+        }
+        if (sb) {
+            try {
+                const sbAccountInfo = sbAccountsInfo.get(reserveAddress);
+                if (sbAccountInfo) {
+                    const price = await getPriceFromSb(sbAccountInfo, sbv2Program);
+                    prices.set(reserveAddress, price);
+                    continue;
+                }
+            } catch { }
+        }
+        prices.set(reserveAddress, 0);
+    }
+
+    return prices;
+};
+
+const getPythAccountsInfo = async (oracles: Map<string, { pyth: string, sb: string }>) => {
     const accountsInfo = new Map<string, AccountInfo<Buffer> | null>();
     const promises = new Set<Promise<void>>();
 
@@ -37,18 +68,7 @@ export const getPythAccountsInfo = async (oracles: Map<string, { pyth: string, s
     return accountsInfo;
 };
 
-export const getPriceFromPyth = (accountInfo: AccountInfo<Buffer> | null) => {
-    if (!accountInfo) {
-        throw new Error("Failed to get price account info");
-    }
-    const priceData = parsePriceData(accountInfo.data);
-    if (!priceData || !priceData.price) {
-        throw new Error("Failed to parse price data");
-    }
-    return priceData.price;
-};
-
-export const getSbAccountsInfo = async (oracles: Map<string, { pyth: string, sb: string }>) => {
+const getSbAccountsInfo = async (oracles: Map<string, { pyth: string, sb: string }>) => {
     const accountsInfo = new Map<string, AccountInfo<Buffer> | null>();
     const promises = new Set<Promise<void>>();
 
@@ -64,7 +84,20 @@ export const getSbAccountsInfo = async (oracles: Map<string, { pyth: string, sb:
     return accountsInfo;
 };
 
-export const getPriceFromSb = async (accountInfo: AccountInfo<Buffer> | null, sbv2: SwitchboardProgram) => {
+const getPriceFromPyth = (accountInfo: AccountInfo<Buffer> | null) => {
+    if (!accountInfo) {
+        throw new Error("Failed to get price account info");
+    }
+    const priceData = parsePriceData(accountInfo.data);
+    if (!priceData || !priceData.price) {
+        throw new Error("Failed to parse price data");
+    }
+    return priceData.price;
+};
+
+
+
+const getPriceFromSb = async (accountInfo: AccountInfo<Buffer> | null, sbv2: SwitchboardProgram) => {
     if (!accountInfo) {
         throw new Error("Failed to get price account info");
     }
